restructure the facts util
