+++ plugins/module_utils/facts/_vm.py
@@ -1,0 +1,333 @@
+# -*- coding: utf-8 -*-
+
+# Copyright: (c) 2023, Ansible Eco Content Team (github.com/eco-ansible-content)
+# GNU General Public License v3.0+ (see LICENSES/GPL-3.0-or-later.txt or https://www.gnu.org/licenses/gpl-3.0.txt)
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+# Note: This utility is considered private, and can only be referenced from inside the vmware.vmware collection.
+#       It may be made public at a later date
+
+from __future__ import absolute_import, division, print_function
+__metaclass__ = type
+
+import base64
+import json
+import os
+
+try:
+    from pyVmomi import vim, VmomiJSONEncoder
+except ImportError:
+    pass
+
+from .._folder_paths import get_folder_path_of_vsphere_object
+
+
+class VmFacts():
+    def __init__(self, vm):
+        self.vm = vm
+
+    def hw_all_facts(self):
+        '''
+        Returns a combined set of all 'hw_' facts
+        '''
+        return {
+            **self.hw_general_facts(),
+            **self.hw_folder_facts(),
+            **self.hw_files_facts(),
+            **self.hw_datastore_facts(),
+            **self.hw_runtime_facts(),
+            **self.hw_network_device_facts()
+        }
+
+    def all_facts(self, content):
+        return {
+            **self.hw_all_facts(),
+            **self.identifier_facts(),
+            **self.custom_attribute_facts(content),
+            **self.advanced_settings_facts(),
+            **self.guest_facts(),
+            **self.ip_facts(),
+            **self.snapshot_facts(),
+            **self.vnc_facts(),
+            **self.tpm_facts()
+        }
+
+    def identifier_facts(self):
+        return {
+            'instance_uuid': self.vm.config.instanceUuid,
+            'moid': self.vm._moId,
+            'vimref': "vim.VirtualMachine:%s" % self.vm._moId
+        }
+
+    def custom_attribute_facts(self, content):
+        custom_value_facts = {}
+        custom_fields_manager = content.customFieldsManager
+
+        for custom_value_obj in self.vm.summary.customValue:
+            if custom_fields_manager is None or not custom_fields_manager.field:
+                custom_value_facts[custom_value_obj.key] = custom_value_obj.value
+                continue
+
+            for field in custom_fields_manager.field:
+                if field.key == custom_value_obj.key:
+                    custom_value_facts[field.name] = custom_value_obj.value
+                    break
+
+        return {
+            'customvalues': custom_value_facts,
+            'annotation': self.vm.config.annotation
+        }
+
+    def advanced_settings_facts(self):
+        output = {}
+        for advanced_setting in self.vm.config.extraConfig:
+            output[advanced_setting.key] = advanced_setting.value
+
+        return {'advanced_settings': output}
+
+    def guest_facts(self):
+        return {
+            'guest_tools_status': get_vm_prop_or_none(self.vm, ('guest', 'toolsRunningStatus')),
+            'guest_tools_version': get_vm_prop_or_none(self.vm, ('guest', 'toolsVersion')),
+            'guest_question': json.loads(json.dumps(self.vm.summary.runtime.question, cls=VmomiJSONEncoder.VmomiJSONEncoder,
+                                                    sort_keys=True, strip_dynamic=True)),
+            'guest_consolidation_needed': self.vm.summary.runtime.consolidationNeeded,
+        }
+
+    def ip_facts(self):
+        facts = {
+            'ipv6': None,
+            'ipv4': None
+        }
+
+        if self.vm.guest.ipAddress:
+            if ':' in self.vm.guest.ipAddress:
+                facts['ipv6'] = self.vm.guest.ipAddress
+            else:
+                facts['ipv4'] = self.vm.guest.ipAddress
+
+        return facts
+
+    def snapshot_facts(self):
+        snapshots = []
+        current_snapshot = None
+        snapshot_facts = list_snapshots(self.vm)
+        if 'snapshots' in snapshot_facts:
+            snapshots = snapshot_facts['snapshots']
+            current_snapshot = snapshot_facts['current_snapshot']
+
+        return {
+            'snapshots': snapshots,
+            'current_snapshot': current_snapshot
+        }
+
+    def vnc_facts(self):
+        facts = {}
+        for opts in self.vm.config.extraConfig:
+            for optkeyname in ['enabled', 'ip', 'port', 'password']:
+                if opts.key.lower() == "remotedisplay.vnc." + optkeyname:
+                    facts[optkeyname] = opts.value
+
+        return {'vnc': facts}
+
+    def tpm_facts(self):
+        facts = {
+            'tpm_present': None,
+            'provider_id': None
+        }
+
+        if hasattr(self.vm.summary.config, 'tpmPresent'):
+            facts['tpm_present'] = self.vm.summary.config.tpmPresent
+
+        if self.vm.config.keyId:
+            facts['provider_id'] = self.vm.config.keyId.providerId.id
+
+        return {'tpm_info': facts}
+
+    def hw_general_facts(self):
+        '''
+        Returns overview and summary 'hw_' facts
+        '''
+        return {
+            'module_hw': True,
+            'hw_name': self.vm.config.name,
+            'hw_power_status': self.vm.summary.runtime.powerState,
+            'hw_guest_full_name': self.vm.summary.guest.guestFullName,
+            'hw_guest_id': self.vm.summary.guest.guestId,
+            'hw_product_uuid': self.vm.config.uuid,
+            'hw_processor_count': self.vm.config.hardware.numCPU,
+            'hw_cores_per_socket': self.vm.config.hardware.numCoresPerSocket,
+            'hw_memtotal_mb': self.vm.config.hardware.memoryMB,
+            'hw_is_template': self.vm.config.template,
+            'hw_version': self.vm.config.version,
+        }
+
+    def hw_folder_facts(self):
+        try:
+            hw_folder = get_folder_path_of_vsphere_object(self.vm)
+        except Exception:
+            hw_folder = None
+
+        return {'hw_folder': hw_folder}
+
+    def hw_files_facts(self):
+        hw_files = []
+        try:
+            files = self.vm.config.files
+            layout = self.vm.layout
+            if files:
+                hw_files = [files.vmPathName]
+                for item in layout.snapshot:
+                    for snap in item.snapshotFile:
+                        if 'vmsn' in snap:
+                            hw_files.append(snap)
+                for item in layout.configFile:
+                    hw_files.append(os.path.join(os.path.dirname(files.vmPathName), item))
+                for item in self.vm.layout.logFile:
+                    hw_files.append(os.path.join(files.logDirectory, item))
+                for item in self.vm.layout.disk:
+                    for disk in item.diskFile:
+                        hw_files.append(disk)
+        except Exception:
+            pass
+
+        return {'hw_files': hw_files}
+
+    def hw_datastore_facts(self):
+        output = []
+        for ds in self.vm.datastore:
+            output.append(ds.info.name)
+
+        return {'hw_datastores': output}
+
+    def hw_runtime_facts(self):
+        output = {
+            'hw_esxi_host': None,
+            'hw_guest_ha_state': None
+        }
+
+        # facts that may or may not exist
+        if self.vm.summary.runtime.host:
+            try:
+                host = self.vm.summary.runtime.host
+                output['hw_esxi_host'] = host.summary.config.name
+                if host.parent and isinstance(host.parent, vim.ClusterComputeResource):
+                    output['hw_cluster'] = host.parent.name
+
+            except vim.fault.NoPermission:
+                # User does not have read permission for the host system,
+                # proceed without this value. This value does not contribute or hamper
+                # provisioning or power management operations.
+                pass
+
+        if self.vm.summary.runtime.dasVmProtection:
+            output['hw_guest_ha_state'] = self.vm.summary.runtime.dasVmProtection.dasProtected
+
+        return output
+
+    def hw_network_device_facts(self):
+        facts = {}
+        hw_interfaces_facts = []
+        ethernet_idx = 0
+
+        vmnet = get_vm_prop_or_none(self.vm, ('guest', 'net'))
+        net_dict = {}
+        if vmnet:
+            for device in vmnet:
+                if device.deviceConfigId > 0:
+                    net_dict[device.macAddress] = list(device.ipAddress)
+
+        for hardware_device in self.vm.config.hardware.device:
+            if not hasattr(hardware_device, 'macAddress'):
+                continue
+
+            if hardware_device.macAddress:
+                mac_addr = hardware_device.macAddress
+                mac_addr_dash = mac_addr.replace(':', '-')
+            else:
+                mac_addr = mac_addr_dash = None
+
+            if (
+                hasattr(hardware_device, "backing")
+                and hasattr(hardware_device.backing, "port")
+                and hasattr(hardware_device.backing.port, "portKey")
+                and hasattr(hardware_device.backing.port, "portgroupKey")
+            ):
+                port_group_key = hardware_device.backing.port.portgroupKey
+                port_key = hardware_device.backing.port.portKey
+            else:
+                port_group_key = None
+                port_key = None
+
+            facts['hw_eth' + str(ethernet_idx)] = {
+                'addresstype': hardware_device.addressType,
+                'label': hardware_device.deviceInfo.label,
+                'macaddress': mac_addr,
+                'ipaddresses': net_dict.get(hardware_device.macAddress, None),
+                'macaddress_dash': mac_addr_dash,
+                'summary': hardware_device.deviceInfo.summary,
+                'portgroup_portkey': port_key,
+                'portgroup_key': port_group_key,
+            }
+            hw_interfaces_facts.append('eth' + str(ethernet_idx))
+            ethernet_idx += 1
+
+        return {
+            'hw_interfaces': hw_interfaces_facts,
+            **facts
+        }
+
+
+def get_vm_prop_or_none(vm, attributes):
+    """Safely get a property or return None"""
+    result = vm
+    for attribute in attributes:
+        try:
+            result = getattr(result, attribute)
+        except (AttributeError, IndexError):
+            return None
+    return result
+
+
+def deserialize_snapshot_obj(obj):
+    return {'id': obj.id,
+            'name': obj.name,
+            'description': obj.description,
+            'creation_time': obj.createTime,
+            'state': obj.state,
+            'quiesced': obj.quiesced}
+
+
+def list_snapshots_recursively(snapshots):
+    snapshot_data = []
+    for snapshot in snapshots:
+        snapshot_data.append(deserialize_snapshot_obj(snapshot))
+        snapshot_data = snapshot_data + list_snapshots_recursively(snapshot.childSnapshotList)
+    return snapshot_data
+
+
+def get_current_snap_obj(snapshots, snapob):
+    snap_obj = []
+    for snapshot in snapshots:
+        if snapshot.snapshot == snapob:
+            snap_obj.append(snapshot)
+        snap_obj = snap_obj + get_current_snap_obj(snapshot.childSnapshotList, snapob)
+    return snap_obj
+
+
+def list_snapshots(vm):
+    result = {}
+    snapshot = get_vm_prop_or_none(vm, ('snapshot',))
+    if not snapshot:
+        return result
+    if vm.snapshot is None:
+        return result
+
+    result['snapshots'] = list_snapshots_recursively(vm.snapshot.rootSnapshotList)
+    current_snapref = vm.snapshot.currentSnapshot
+    current_snap_obj = get_current_snap_obj(vm.snapshot.rootSnapshotList, current_snapref)
+    if current_snap_obj:
+        result['current_snapshot'] = deserialize_snapshot_obj(current_snap_obj[0])
+    else:
+        result['current_snapshot'] = dict()
+    return result
